신규 및 수정모드인것과 계산용 위젯을 구분하여 보이도록 함. 자동 계산 기능 구현됨.

+++ lib/model/ingredient.dart
@@ -6,7 +6,7 @@ class Ingredient extends HiveObject {
   @HiveField(0)
   String name;
   @HiveField(1)
-  int count;
+  double count;
   @HiveField(2)
   String unit;
